backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-dotenv>=1.0.0
pillow>=9.5.0
google-generativeai>=0.3.1
pydantic>=2.0.0